Fix setting the paper size. closes #39
If you allow decimal numbers as input you must parse them as decimal
numbers. ;)
Also binding the size of the SnapGrid to the drawPanel helps displaying
the actual size of the drawPanel. ;)

+++ src/main/scala/de/thm/move/controllers/MoveCtrl.scala
@@ -392,8 +392,8 @@ class MoveCtrl extends Initializable {
     strOpt.flatMap { x =>
       try {
         val ar = x.split(";")
-        val width = ar(0).toInt
-        val height = ar(1).toInt
+        val width = ar(0).toDouble
+        val height = ar(1).toDouble
         Some((width,height))
       } catch {
         case _:NumberFormatException => None
